+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+* Add WeakAuras EVENTS handler. Fire the lib events in Weakaura if the addon is loaded.
+
 ## [2.0.10] - 2024-05-14
 
 ### Fixed